renamed source files to include manager
