+++ src/index.ts
@@ -1,4 +1,4 @@
 /*!
  * Copyright (c) 2022 Digital Credentials Consortium. All rights reserved.
  */
-export { createStatusListManager } from './credential-status-index';
+export { createStatusListManager } from './credential-status-manager-index';
